UI: Added permission system for mobile center and oauth clients

+++ ui-ngx/src/app/modules/home/pages/mobile/applications/mobile-app-table-config.resolver.ts
@@ -35,6 +35,7 @@ import {
   CellActionDescriptor,
   CellActionDescriptorType,
   DateEntityTableColumn,
+  defaultEntityTablePermissions,
   EntityTableColumn,
   EntityTableConfig
 } from '@home/models/entity/entities-table-config.models';
@@ -58,6 +59,8 @@ import {
   MobileAppDeleteDialogData,
   RemoveAppDialogComponent
 } from '@home/pages/mobile/applications/remove-app-dialog.component';
+import { Operation, Resource } from '@shared/models/security.models';
+import { UserPermissionsService } from '@core/http/user-permissions.service';
 
 @Injectable()
 export class MobileAppTableConfigResolver  {
@@ -68,7 +71,8 @@ export class MobileAppTableConfigResolver  {
               private datePipe: DatePipe,
               private mobileAppService: MobileAppService,
               private truncatePipe: TruncatePipe,
-              private dialog: MatDialog
+              private dialog: MatDialog,
+              private userPermissionsService: UserPermissionsService,
               ) {
     this.config.selectionEnabled = false;
     this.config.entityType = EntityType.MOBILE_APP;
@@ -135,18 +139,22 @@ export class MobileAppTableConfigResolver  {
   }
 
   resolve(_route: ActivatedRouteSnapshot): EntityTableConfig<MobileApp> {
+    defaultEntityTablePermissions(this.userPermissionsService, this.config);
     return this.config;
   }
 
   private configureCellActions(): Array<CellActionDescriptor<MobileApp>> {
-    return [
-      {
-        name: this.translate.instant('action.delete'),
-        icon: 'delete',
-        isEnabled: () => true,
-        onAction: ($event, entity) => this.deleteEntity($event, entity)
-      }
-    ];
+    if (this.userPermissionsService.hasGenericPermission(Resource.MOBILE_APP, Operation.DELETE)) {
+      return [
+        {
+          name: this.translate.instant('action.delete'),
+          icon: 'delete',
+          isEnabled: () => true,
+          onAction: ($event, entity) => this.deleteEntity($event, entity)
+        }
+      ];
+    }
+    return [];
   }
 
   private deleteEntity($event: Event, entity: MobileApp) {
